fix: domain update

+++ src/main/java/com/example/catchtable/domain/OrderStatus.java
@@ -1,7 +1,5 @@
 package com.example.catchtable.domain;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,11 +8,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class OrderStatus {
 
-  private Integer id;
+  private Long id; // int unsigned -> Long
   private String type;
 
   // 정적 팩토리 메소드
-  public static OrderStatus fromEntity(Integer id, String type) {
+  public static OrderStatus fromEntity(Long id, String type) {
     OrderStatus orderStatus = new OrderStatus();
     orderStatus.id = id;
     orderStatus.type = type;
