+++ src/main/java/com/example/catchtable/domain/PickupStatus.java
@@ -1,6 +1,5 @@
 package com.example.catchtable.domain;
 
-import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,14 +8,16 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PickupStatus {
 
-  private Integer id;
+  private Long id; // int unsigned -> Long
   private String type;
 
   // 정적 팩토리 메소드
-  public static PickupStatus fromEntity(final Integer id, final String type) {
+  public static PickupStatus fromEntity(final Long id, final String type) {
     PickupStatus pickupStatus = new PickupStatus();
     pickupStatus.id = id;
     pickupStatus.type = type;
     return pickupStatus;
   }
+
+  // ... (필요한 비즈니스 로직 추가)
 }
