+++ src/main/java/com/example/catchtable/domain/ReservationHistory.java
@@ -2,7 +2,6 @@ package com.example.catchtable.domain;
 
 import static com.example.catchtable.util.LocalDateTimeUtil.toLocalDateTimeOrNull;
 
-import jakarta.validation.constraints.NotNull;
 import java.sql.Timestamp;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -14,16 +13,17 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class ReservationHistory {
 
-  private Integer id;
-  private Integer statusId;
-  private LocalDateTime visitedAt;
-  private Integer reservationId;
-  private LocalDate bookingTime;
-  private Integer guestsCount;
+  private Long id; // int unsigned -> Long
+  private Long statusId; // int unsigned -> Long
+  private LocalDateTime visitedAt; // datetime -> LocalDateTime
+  private Long reservationId; // int unsigned -> Long
+  private LocalDate bookingTime; // date -> LocalDate
+  private Integer guestsCount; // tinyint(6) unsigned -> Integer
 
   // 정적 팩토리 메소드
   public static ReservationHistory fromEntity(
-      final Integer id, final Integer statusId, final Timestamp visitedAt, final Integer reservationId, final LocalDate bookingTime, final Integer guestsCount) {
+      final Long id, final Long statusId, final Timestamp visitedAt, final Long reservationId,
+      final LocalDate bookingTime, final Integer guestsCount) {
     ReservationHistory reservationHistory = new ReservationHistory();
     reservationHistory.id = id;
     reservationHistory.statusId = statusId;
@@ -33,4 +33,6 @@ public class ReservationHistory {
     reservationHistory.guestsCount = guestsCount;
     return reservationHistory;
   }
+
+  // ... (필요한 비즈니스 로직 추가)
 }
